+++ source/voxelman/graphics/font/fontmanager.d
@@ -19,9 +19,12 @@ class FontManager
 {
 	public:
 
+	Font* defaultFont;
+
 	this(TextureAtlas texAtlas)
 	{
 		this.texAtlas = texAtlas;
+		defaultFont = createFont("font_hor.png", 10, GLYPHS);
 	}
 
 	Font* createFont(in string filename, in uint size, in dchar[] chars = GLYPHS)