+++ source/voxelman/graphics/font/package.d
@@ -1,0 +1,12 @@
+/**
+Copyright: Copyright (c) 2017 Andrey Penechko.
+License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
+Authors: Andrey Penechko.
+*/
+
+module voxelman.graphics.font;
+
+public import voxelman.graphics.font.bitmapfontloader;
+public import voxelman.graphics.font.font;
+public import voxelman.graphics.font.fontmanager;
+public import voxelman.graphics.font.textrenderer;