+++ source/voxelman/graphics/package.d
@@ -5,16 +5,23 @@ Authors: Andrey Penechko.
 */
 module voxelman.graphics;
 
+public import derelict.opengl3.gl3;
+
+public import voxelman.graphics.animation;
 public import voxelman.graphics.batch2d;
-public import voxelman.graphics.bitmap;
 public import voxelman.graphics.batch;
+public import voxelman.graphics.bitmap;
+public import voxelman.graphics.bufferrenderer;
 public import voxelman.graphics.color;
+public import voxelman.graphics.font;
 public import voxelman.graphics.fpscamera;
 public import voxelman.graphics.glerrors;
 public import voxelman.graphics.irenderer;
 public import voxelman.graphics.oglrenderer;
+public import voxelman.graphics.resourcemanager;
 public import voxelman.graphics.shaderprogram;
 public import voxelman.graphics.shaders;
+public import voxelman.graphics.sprite;
 public import voxelman.graphics.texture;
 public import voxelman.graphics.textureatlas;
 public import voxelman.graphics.vao;