+++ source/voxelman/platform/package.d
@@ -1,0 +1,16 @@
+/**
+Copyright: Copyright (c) 2017 Andrey Penechko.
+License: $(WEB boost.org/LICENSE_1_0.txt, Boost License 1.0).
+Authors: Andrey Penechko.
+*/
+
+module voxelman.platform;
+
+public import voxelman.platform.compression;
+public import voxelman.platform.glfwwindow;
+public import voxelman.platform.input;
+public import voxelman.platform.isharedcontext;
+public import voxelman.platform.iwindow;
+public import voxelman.platform.keynamemap;
+public import voxelman.platform.monitorman;
+public import voxelman.platform.platform;